// src/generated/contentfulTypes.d.ts
// THIS FILE IS AUTOMATICALLY GENERATED. DO NOT MODIFY IT.

import { Asset, Entry } from "contentful";
import { Document } from "@contentful/rich-text-types";

export interface IAbschnittFields {
  /** Titel */
  titel?: string | undefined;

  /** Cards */
  cards?: ICard[] | undefined;
}

export interface IAbschnitt extends Entry<IAbschnittFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "abschnitt";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface ICardFields {
  /** Title */
  title?: string | undefined;

  /** Text */
  text?: Document | undefined;
}

export interface ICard extends Entry<ICardFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "card";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IPageFields {
  /** Title */
  title?: string | undefined;

  /** Abschnitte */
  abschnitte?: IAbschnitt[] | undefined;
}

export interface IPage extends Entry<IPageFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "page";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export type CONTENT_TYPE = "abschnitt" | "card" | "page";

export type IEntry = IAbschnitt | ICard | IPage;

export type LOCALE_CODE = "en-US";

export type CONTENTFUL_DEFAULT_LOCALE_CODE = "en-US";